Moved JS files to distributions/JavaScript/ Change emcc to use only libgsl.a no need of libgsl.so Added WASM param to compile_to_js.sh to let compile without Used ALL_FUNCTIONS flag in emcc instead of generating list Rewrite instruction to compile to JS JS compiled library now appears in /distributions/JavaScript/output/ Removed unsed libgsl.so
