+++ .gitignore
@@ -3,3 +3,4 @@ cmake-build-debug/
 libat.*
 .idea/*
 *.xml
+/distributions/JavaScript/output/
